+++ packages/proposal/src/api/utils.ts
@@ -3,5 +3,5 @@ import { ExternalApiProposalResponse } from '../models/response';
 import { DEFENDER_APP_URL } from '@openzeppelin/defender-sdk-base-client';
 
 export function getProposalUrl(proposal: Pick<ExternalApiProposalResponse, 'proposalId'>): string {
-  return `${DEFENDER_APP_URL}/v2/#/actions/executable/${proposal.proposalId}`;
+  return `${DEFENDER_APP_URL}/v2/#/transaction-proposals/${proposal.proposalId}`;
 }