+++ packages/relay-signer/src/relayer.ts
@@ -13,6 +13,7 @@ import {
   PaginatedTransactionResponse,
   RelayerTransaction,
   RelayerTransactionPayload,
+  TransactionDeleteResponse,
 } from './models/transactions';
 import { isApiCredentials, isActionCredentials, validatePayload } from './ethers/utils';
 import { RelaySignerClient } from './api';
@@ -120,6 +121,10 @@ export class Relayer implements IRelayer {
   public replaceTransactionByNonce(nonce: number, payload: RelayerTransactionPayload): Promise<RelayerTransaction> {
     validatePayload(payload);
     return this.relayer.replaceTransactionByNonce(nonce, payload);
+  }
+
+  public cancelTransactionById(id: string): Promise<TransactionDeleteResponse> {
+    return this.relayer.cancelTransactionById(id);
   }
 
   public getTransaction(id: string): Promise<RelayerTransaction> {