+++ packages/defender-sdk/CHANGELOG.md
@@ -1,5 +1,86 @@
 # @openzeppelin/defender-sdk
 
+## 1.14.4
+
+### Patch Changes
+
+- 3a1d0f3: feat: Add missing dependencies that caused relay signer to fail.
+  feat: Add support to isProduction flag for private networks.
+  chore: Update transaction proposal route.
+  feat: add support to deployment metadata.
+  feat: add support for cancelling intents.
+- Updated dependencies [3a1d0f3]
+  - @openzeppelin/defender-sdk-notification-channel-client@1.14.4
+  - @openzeppelin/defender-sdk-relay-signer-client@1.14.4
+  - @openzeppelin/defender-sdk-proposal-client@1.14.4
+  - @openzeppelin/defender-sdk-account-client@1.14.4
+  - @openzeppelin/defender-sdk-key-value-store-client@1.14.4
+  - @openzeppelin/defender-sdk-monitor-client@1.14.4
+  - @openzeppelin/defender-sdk-network-client@1.14.4
+  - @openzeppelin/defender-sdk-action-client@1.14.4
+  - @openzeppelin/defender-sdk-deploy-client@1.14.4
+  - @openzeppelin/defender-sdk-relay-client@1.14.4
+  - @openzeppelin/defender-sdk-base-client@1.14.4
+
+## 1.14.3
+
+### Patch Changes
+
+- 19cd7a9: feat: Add relayers usage limiting
+  feat: Add an example contract call
+- Updated dependencies [19cd7a9]
+  - @openzeppelin/defender-sdk-account-client@1.14.3
+  - @openzeppelin/defender-sdk-action-client@1.14.3
+  - @openzeppelin/defender-sdk-base-client@1.14.3
+  - @openzeppelin/defender-sdk-deploy-client@1.14.3
+  - @openzeppelin/defender-sdk-monitor-client@1.14.3
+  - @openzeppelin/defender-sdk-network-client@1.14.3
+  - @openzeppelin/defender-sdk-notification-channel-client@1.14.3
+  - @openzeppelin/defender-sdk-proposal-client@1.14.3
+  - @openzeppelin/defender-sdk-relay-client@1.14.3
+  - @openzeppelin/defender-sdk-relay-signer-client@1.14.3
+
+## 1.14.1
+
+### Patch Changes
+
+- 845d942: chore: fix vulnerabilities
+  chore: Add utility for verifying webhooks signatures
+  chore: Optimise api access by default
+- Updated dependencies [845d942]
+  - @openzeppelin/defender-sdk-account-client@1.14.1
+  - @openzeppelin/defender-sdk-action-client@1.14.1
+  - @openzeppelin/defender-sdk-base-client@1.14.1
+  - @openzeppelin/defender-sdk-deploy-client@1.14.1
+  - @openzeppelin/defender-sdk-monitor-client@1.14.1
+  - @openzeppelin/defender-sdk-network-client@1.14.1
+  - @openzeppelin/defender-sdk-notification-channel-client@1.14.1
+  - @openzeppelin/defender-sdk-proposal-client@1.14.1
+  - @openzeppelin/defender-sdk-relay-client@1.14.1
+  - @openzeppelin/defender-sdk-relay-signer-client@1.14.1
+
+## 1.14.0
+
+### Minor Changes
+
+- 164af52: chore: fix vulnerabilities
+  chore: Add utility for verifying webhooks signatures
+  chore: Optimise api access by default
+
+### Patch Changes
+
+- Updated dependencies [164af52]
+  - @openzeppelin/defender-sdk-account-client@1.14.0
+  - @openzeppelin/defender-sdk-action-client@1.14.0
+  - @openzeppelin/defender-sdk-base-client@1.14.0
+  - @openzeppelin/defender-sdk-deploy-client@1.14.0
+  - @openzeppelin/defender-sdk-monitor-client@1.14.0
+  - @openzeppelin/defender-sdk-network-client@1.14.0
+  - @openzeppelin/defender-sdk-notification-channel-client@1.14.0
+  - @openzeppelin/defender-sdk-proposal-client@1.14.0
+  - @openzeppelin/defender-sdk-relay-client@1.14.0
+  - @openzeppelin/defender-sdk-relay-signer-client@1.14.0
+
 ## 1.13.4
 
 ### Patch Changes